+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Django Field Audit change log
 
+## v1.2.2 - 2022-08-19
+
+- Add management command for bootstrapping
+- Add new "top up" bootstrap action
+
 ## v1.2.1 - 2022-08-12
 
 - Fix migration constraint condition ordering which caused Django<4.0 to think